+++ packages/press-ui/src/packages/press-dialog-plus/handler-helper.d.ts
@@ -1,1 +1,3 @@
-export declare function addFunctionForDialog({ Dialog, queue, currentOptions, defaultOptions, updateQueue, updateCurrentOptions }: any): any;
+export declare function addFunctionForDialog({
+  Dialog, queue, currentOptions, defaultOptions, updateQueue, updateCurrentOptions,
+}: any): any;