+++ packages/press-ui/src/packages/press-dialog-plus/handler.d.ts
@@ -1,7 +1,14 @@
 interface IDialog {
   (options: any): any;
   show: (options: any) => Promise<any>;
-  clear: () => any;
+  confirm: (options: any) => Promise<any>;
+  alert: (options: any) => Promise<any>;
+
+  clear: () => void;
+  close: () => void;
+
+  setDefaultOptions: (options: any) => void;
+  resetDefaultOptions: () => void;
 }
 const Dialog: IDialog;